+++ LibClassicSwingTimerAPI.lua
@@ -26,8 +26,8 @@ local ranged_swing = nil
 local reset_ranged_swing = nil
 
 local Unit = {
+	id = nil,
 	GUID = nil,
-	type = nil,
 	class = nil,
 
 	mainSpeed = 0,
@@ -72,7 +72,7 @@ end
 
 function Unit:CalculateDelta()
 	if self.offSpeed > 0 and self.mainExpirationTime ~= nil and self.offExpirationTime ~= nil then
-		self.callbacks:Fire("SWING_TIMER_DELTA", self.mainExpirationTime - self.offExpirationTime, self.type)
+		self.callbacks:Fire("UNIT_SWING_TIMER_DELTA", self.id, self.mainExpirationTime - self.offExpirationTime)
 	end
 end
 
@@ -82,10 +82,10 @@ function Unit:SwingStart(hand, startTime, isReset)
 			self.mainTimer:Cancel()
 		end
 		self.lastMainSwing = startTime
-		local mainSpeed, _ = UnitAttackSpeed(self.type)
+		local mainSpeed, _ = UnitAttackSpeed(self.id)
 		self.mainSpeed = mainSpeed
 		self.mainExpirationTime = self.lastMainSwing + self.mainSpeed
-		self.callbacks:Fire("SWING_TIMER_START", self.mainSpeed, self.mainExpirationTime, hand, self.type)
+		self.callbacks:Fire("UNIT_SWING_TIMER_START", self.id, self.mainSpeed, self.mainExpirationTime, hand)
 		if self.mainSpeed > 0 and self.mainExpirationTime - GetTime() > 0 then
 			self.mainTimer = C_Timer.NewTimer(self.mainExpirationTime - GetTime(), function()
 				self:SwingEnd("mainhand")
@@ -96,7 +96,7 @@ function Unit:SwingStart(hand, startTime, isReset)
 			self.offTimer:Cancel()
 		end
 		self.lastOffSwing = startTime
-		local _, offSpeed = UnitAttackSpeed(self.type)
+		local _, offSpeed = UnitAttackSpeed(self.id)
 		self.offSpeed = offSpeed or 0
 		self.offExpirationTime = self.lastOffSwing + self.offSpeed
 		if self.calculaDeltaTimer then
@@ -105,9 +105,9 @@ function Unit:SwingStart(hand, startTime, isReset)
 		if self.offSpeed > 0 and self.firstOffSwing == false and self.isAttacking then
 			self.offExpirationTime = self.lastOffSwing + (self.offSpeed / 2)
 			self:CalculateDelta()
-			self.callbacks:Fire("SWING_TIMER_UPDATE", self.offSpeed, self.offExpirationTime, hand, self.type)
+			self.callbacks:Fire("UNIT_SWING_TIMER_UPDATE", self.id, self.offSpeed, self.offExpirationTime, hand)
 		elseif self.offSpeed > 0 then
-			self.callbacks:Fire("SWING_TIMER_START", self.offSpeed, self.offExpirationTime, hand, self.type)
+			self.callbacks:Fire("UNIT_SWING_TIMER_START", self.id, self.offSpeed, self.offExpirationTime, hand)
 			self.calculaDeltaTimer = C_Timer.NewTimer(self.offSpeed / 2, function()
 				self:CalculateDelta()
 			end)
@@ -126,7 +126,7 @@ function Unit:SwingStart(hand, startTime, isReset)
 			self.rangedSpeed = self.rangedSpeed
 			self.lastRangedSwing = startTime
 			self.rangedExpirationTime = self.lastRangedSwing + self.rangedSpeed
-			self.callbacks:Fire("SWING_TIMER_START", self.rangedSpeed, self.rangedExpirationTime, hand, self.type)
+			self.callbacks:Fire("UNIT_SWING_TIMER_START", self.id, self.rangedSpeed, self.rangedExpirationTime, hand)
 			if self.rangedExpirationTime - GetTime() > 0 then
 				self.rangedTimer = C_Timer.NewTimer(self.rangedExpirationTime - GetTime(), function()
 					self:SwingEnd("ranged")
@@ -137,15 +137,15 @@ function Unit:SwingStart(hand, startTime, isReset)
 end
 
 function Unit:SwingEnd(hand)
-	self.callbacks:Fire("SWING_TIMER_STOP", hand, self.type)
+	self.callbacks:Fire("UNIT_SWING_TIMER_STOP", self.id, hand)
 	if (self.casting or self.channeling) and self.isAttacking and hand ~= "ranged" then
 		local now = GetTime()
 		if isRetail and hand == "mainhand" then		
 			self:SwingStart(hand, now, true)
-			self.callbacks:Fire("SWING_TIMER_CLIPPED", hand, self.type)
+			self.callbacks:Fire("UNIT_SWING_TIMER_CLIPPED", self.id, hand)
 		elseif isClassicOrBCCOrWrath then
 			self:SwingStart(hand, now, true)
-			self.callbacks:Fire("SWING_TIMER_CLIPPED", hand, self.type)
+			self.callbacks:Fire("UNIT_SWING_TIMER_CLIPPED", self.id, hand)
 		end
 	end
 end
@@ -153,17 +153,27 @@ end
 lib.callbacks = lib.callbacks or LibStub("CallbackHandler-1.0"):New(lib)
 
 function lib:getUnit(unit)
-	if self.player.GUID == unit or self.player.type == unit then
+	if self.player.GUID == unit or self.player.id == unit then
 		return self.player
-	elseif self.target.GUID == unit  or self.player.type == unit then
+	elseif self.target.GUID == unit  or self.player.id == unit then
 		return self.target
 	else
 		return nil
 	end
 end
 
-function lib:SwingTimerInfo(hand, unitType)
-	local unit = lib:getUnit(unitType)
+function lib:SwingTimerInfo(hand)
+	if hand == "mainhand" then
+		return self.player.mainSpeed, self.player.mainExpirationTime, self.player.lastMainSwing
+	elseif hand == "offhand" then
+		return self.player.offSpeed, self.player.offExpirationTime, self.player.lastOffSwing
+	elseif hand == "ranged" then
+		return self.player.rangedSpeed, self.player.rangedExpirationTime, self.player.lastRangedSwing
+	end
+end
+
+function lib:UnitSwingTimerInfo(unitId, hand)
+	local unit = lib:getUnit(unitId)
 	if not unit then
 		return
 	end
@@ -181,9 +191,9 @@ function lib:ADDON_LOADED(_, addOnName)
 		return
 	end
 
-	self.player = Unit:new({type="player"})
+	self.player = Unit:new({id="player"})
 	self.player.callbacks = self.callbacks
-	self.target = Unit:new({type="target", class="TARGET"})
+	self.target = Unit:new({id="target", class="TARGET"})
 	self.target.callbacks = self.callbacks
 end
 
@@ -314,7 +324,7 @@ function lib:COMBAT_LOG_EVENT_UNFILTERED(_, ts, subEvent, _, sourceGUID, _, _, _
 		else
 			unit.mainExpirationTime = swing_timer_reduced_40p
 		end
-		self.callbacks:Fire("SWING_TIMER_UPDATE", unit.mainSpeed, unit.mainExpirationTime, "mainhand", unit.type)
+		self.callbacks:Fire("UNIT_SWING_TIMER_UPDATE", unit.id, unit.mainSpeed, unit.mainExpirationTime, "mainhand")
 		if unit.mainSpeed > 0 and unit.mainExpirationTime - GetTime() > 0 then
 			unit.mainTimer = C_Timer.NewTimer(unit.mainExpirationTime - GetTime(), function()
 				unit:SwingEnd("mainhand")
@@ -363,14 +373,14 @@ function lib:UNIT_ATTACK_SPEED(unitGUID)
 	if unit.offTimer then
 		unit.offTimer:Cancel()
 	end
-	local mainSpeedNew, offSpeedNew = UnitAttackSpeed(unit.type)
+	local mainSpeedNew, offSpeedNew = UnitAttackSpeed(unit.id)
 	offSpeedNew = offSpeedNew or 0
 	if mainSpeedNew > 0 and unit.mainSpeed > 0 and mainSpeedNew ~= unit.mainSpeed then
 		local multiplier = mainSpeedNew / unit.mainSpeed
 		local timeLeft = (unit.lastMainSwing + unit.mainSpeed - now) * multiplier
 		unit.mainSpeed = mainSpeedNew
 		unit.mainExpirationTime = now + timeLeft
-		self.callbacks:Fire("SWING_TIMER_UPDATE", unit.mainSpeed, unit.mainExpirationTime, "mainhand", unit.type)
+		self.callbacks:Fire("UNIT_SWING_TIMER_UPDATE", unit.id, unit.mainSpeed, unit.mainExpirationTime, "mainhand")
 		if unit.mainSpeed > 0 and unit.mainExpirationTime - GetTime() > 0 then
 			unit.mainTimer = C_Timer.NewTimer(unit.mainExpirationTime - GetTime(), function()
 				unit:SwingEnd("mainhand")
@@ -385,7 +395,7 @@ function lib:UNIT_ATTACK_SPEED(unitGUID)
 		if unit.calculaDeltaTimer ~= nil then
 			unit.calculaDeltaTimer:Cancel()
 		end
-		self.callbacks:Fire("SWING_TIMER_UPDATE", unit.offSpeed, unit.offExpirationTime, "offhand", unit.type)
+		self.callbacks:Fire("UNIT_SWING_TIMER_UPDATE", unit.id, unit.offSpeed, unit.offExpirationTime, "offhand")
 		if unit.offSpeed > 0 and unit.offExpirationTime - GetTime() > 0 then
 			unit.offTimer = C_Timer.NewTimer(unit.offExpirationTime - GetTime(), function()
 				unit:SwingEnd("offhand")
@@ -406,7 +416,7 @@ function lib:UNIT_SPELLCAST_INTERRUPTED_OR_FAILED(_, unitType, _, spell)
 			if unit.mainExpirationTime < GetTime() and unit.isAttacking then
 				unit.mainExpirationTime = unit.mainExpirationTime + unit.mainSpeed
 			end
-			self.callbacks:Fire("SWING_TIMER_UPDATE", unit.mainSpeed, unit.mainExpirationTime, "mainhand", unit.type)
+			self.callbacks:Fire("UNIT_SWING_TIMER_UPDATE", unit.id, unit.mainSpeed, unit.mainExpirationTime, "mainhand")
 			unit.mainTimer = C_Timer.NewTimer(unit.mainExpirationTime - GetTime(), function()
 				self:SwingEnd("mainhand")
 			end)
@@ -415,7 +425,7 @@ function lib:UNIT_SPELLCAST_INTERRUPTED_OR_FAILED(_, unitType, _, spell)
 			if unit.offExpirationTime < GetTime() and unit.isAttacking then
 				unit.offExpirationTime = unit.offExpirationTime + unit.offSpeed
 			end
-			self.callbacks:Fire("SWING_TIMER_UPDATE", unit.offSpeed, unit.offExpirationTime, "offhand", unit.type)
+			self.callbacks:Fire("UNIT_SWING_TIMER_UPDATE", unit.id, unit.offSpeed, unit.offExpirationTime, "offhand")
 			unit.offTimer = C_Timer.NewTimer(unit.offExpirationTime - GetTime(), function()
 				self:SwingEnd("offhand")
 			end)
@@ -461,14 +471,14 @@ function lib:UNIT_SPELLCAST_SUCCEEDED(_, unitType, _, spell)
 		unit.pauseSwingTime = nil
 		if unit.mainSpeed > 0 then
 			unit.mainExpirationTime = unit.mainExpirationTime + offset
-			self.callbacks:Fire("SWING_TIMER_UPDATE", unit.mainSpeed, unit.mainExpirationTime, "mainhand", unit.type)
+			self.callbacks:Fire("UNIT_SWING_TIMER_UPDATE", unit.id, unit.mainSpeed, unit.mainExpirationTime, "mainhand")
 			unit.mainTimer = C_Timer.NewTimer(unit.mainExpirationTime - now, function()
 				self:SwingEnd("mainhand")
 			end)
 		end
 		if unit.offSpeed > 0 then
 			unit.offExpirationTime = unit.offExpirationTime + offset
-			self.callbacks:Fire("SWING_TIMER_UPDATE", unit.offSpeed, unit.offExpirationTime, "offhand", unit.type)
+			self.callbacks:Fire("UNIT_SWING_TIMER_UPDATE", unit.id, unit.offSpeed, unit.offExpirationTime, "offhand")
 			unit.offTimer = C_Timer.NewTimer(unit.offExpirationTime - now, function()
 				self:SwingEnd("offhand")
 			end)
@@ -508,13 +518,13 @@ function lib:UNIT_SPELLCAST_START(_, unitType, _, spell)
 		if spell and pause_swing_spells[spell] then
 			unit.pauseSwingTime = now
 			if unit.mainSpeed > 0 then
-				self.callbacks:Fire("SWING_TIMER_PAUSED", "mainhand", unit.type)
+				self.callbacks:Fire("UNIT_SWING_TIMER_PAUSED", unit.id, "mainhand")
 				if unit.mainTimer then
 					unit.mainTimer:Cancel()
 				end
 			end
 			if unit.offSpeed > 0 then
-				self.callbacks:Fire("SWING_TIMER_PAUSED", "offhand", unit.type)
+				self.callbacks:Fire("UNIT_SWING_TIMER_PAUSED", unit.id, "offhand")
 				if unit.offTimer then
 					unit.offTimer:Cancel()
 				end
@@ -602,6 +612,23 @@ frame:SetScript("OnEvent", function(_, event, ...)
 		lib[event](lib, event, ...)
 	end
 end)
+
+--[[
+	Backward compatibility continue to fire EVENTS with SWING_TIMER_ format for player unit.
+]]--
+local EventBackwardCompatibility = function(event, ...)
+	local unitId = select(1,...)
+	if unitId == "player" then
+		lib.callbacks:Fire(string.gsub(event,"UNIT_",""), select(2,...))
+	end
+end
+
+lib.RegisterCallback(lib, "UNIT_SWING_TIMER_START", EventBackwardCompatibility)
+lib.RegisterCallback(lib, "UNIT_SWING_TIMER_UPDATE", EventBackwardCompatibility)
+lib.RegisterCallback(lib, "UNIT_SWING_TIMER_CLIPPED", EventBackwardCompatibility)
+lib.RegisterCallback(lib, "UNIT_SWING_TIMER_PAUSED", EventBackwardCompatibility)
+lib.RegisterCallback(lib, "UNIT_SWING_TIMER_STOP", EventBackwardCompatibility)
+lib.RegisterCallback(lib, "UNIT_SWING_TIMER_DELTA", EventBackwardCompatibility)
 
 --[[
 	Set table data based on current game version